Refactor and Cleanup
Allows tests to run on netstandard by moving the logoptions to a separate project.
