+++ AppInspector.CLI/AppInspector.CLI.csproj
@@ -40,6 +40,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\AppInspector.Logging\AppInspector.Logging.csproj" />
     <ProjectReference Include="..\AppInspector\AppInspector.Commands.csproj" />
     <ProjectReference Include="..\RulesEngine\AppInspector.RulesEngine.csproj" />
   </ItemGroup>
