+++ AppInspector.CLI/CLICmdOptions.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (C) Microsoft. All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
+using Microsoft.ApplicationInspector.Logging;
+
 namespace Microsoft.ApplicationInspector.CLI
 {
     using CommandLine;
