+++ AppInspector.CLI/Program.cs
@@ -1,6 +1,8 @@
 ﻿//Copyright(c) Microsoft Corporation.All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.ApplicationInspector.Logging;
+
 namespace Microsoft.ApplicationInspector.CLI
 {
     using CommandLine;
