+++ AppInspector.CLI/Writers/JsonWriter.cs
@@ -1,7 +1,6 @@
 ﻿namespace Microsoft.ApplicationInspector.CLI.Writers
 {
     using Microsoft.ApplicationInspector.Commands;
-    using Microsoft.ApplicationInspector.Common;
     using Microsoft.Extensions.Logging;
     using Microsoft.Extensions.Logging.Abstractions;
     using Newtonsoft.Json;
