+++ AppInspector.CLI/Writers/VerifyRulesTextWriter.cs
@@ -7,7 +7,6 @@ namespace Microsoft.ApplicationInspector.CLI
     using Microsoft.ApplicationInspector.Common;
     using Microsoft.Extensions.Logging;
     using Microsoft.Extensions.Logging.Abstractions;
-    using System;
     using System.IO;
 
     internal class VerifyRulesTextWriter : CommandResultsWriter
