+++ AppInspector.Tests.CLI/Logger/LogOptionsTests.cs
