+++ AppInspector.Tests/RuleProcessor/LoadRulesTests.cs
@@ -1,9 +1,7 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using Microsoft.ApplicationInspector.CLI;
-using Microsoft.ApplicationInspector.Common;
+using Microsoft.ApplicationInspector.Logging;
 using Microsoft.ApplicationInspector.RulesEngine;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
