+++ AppInspector.Tests/TestHelpers.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
-using Microsoft.ApplicationInspector.CLI;
+using Microsoft.ApplicationInspector.Logging;
 using Microsoft.Extensions.Logging;
 
 namespace AppInspector.Tests
